+++ R/data.R
@@ -12,3 +12,19 @@
 #' @references
 #' \insertAllCited{}
 "Anderson2010"
+
+
+#' @title 10 experimental studies from from
+#' \insertCite{Anderson2010;textual}{RoBMA} that meet the best practice criteria
+#'
+#' @description The dataset contains correlational coefficients, sample
+#' sizes, and labels for 27 experimental studies focusing on the effect of
+#' violent videogames on aggressive behavior. The full original data can
+#' found at https://github.com/Joe-Hilgard/Anderson-meta.
+#'
+#'
+#' @format A data.frame with 3 columns and 10 observations.
+#'
+#' @references
+#' \insertAllCited{}
+"Bem2011"