+++ src/weightedt.cc
@@ -3,53 +3,81 @@
 #include "distributions/DWT2.h"
 #include "distributions/DWN1.h"
 #include "distributions/DWN2.h"
-#include "distributions/DTboost.h"
-#include "distributions/DWT1boost.h"
-#include "distributions/DWT2boost.h"
-#include "distributions/DNboost.h"
-#include "distributions/DWN1boost.h"
-#include "distributions/DWN2boost.h"
-#include "functions/LogWeightedtFun.h"
+
+#include "transformations/d.h"
+#include "transformations/r.h"
+#include "transformations/z.h"
+#include "transformations/logOR.h"
 
 namespace jags { 
-namespace weightedt { // module namespace
+  namespace weightedt { // module namespace
 
-  // JAGS module class
-  class WTModule : public Module {
-    public:
-      WTModule();
-      ~WTModule();
-  };
+    // JAGS module class
+    class WTModule : public Module {
+      public:
+        WTModule();
+        ~WTModule();
+    };
 
-  // constructor (executed when loading the module)
-  WTModule::WTModule() : Module("RoBMA"){
-    insert(new DWT1);
-    insert(new DWT2);
-    insert(new DWN1);
-    insert(new DWN2);
-    insert(new DTboost);
-    insert(new DWT1boost);
-    insert(new DWT2boost);
-    insert(new DNboost);
-    insert(new DWN1boost);
-    insert(new DWN2boost);
-    //load functions
-    insert(new LogWeightedtFun);
-  }
-  
-  // destructor (executed when unloading the module)
-  WTModule::~WTModule() {
-    std::vector<Function*> const &fvec = functions();
-    for (unsigned int i = 0; i < fvec.size(); ++i) {
-      delete fvec[i];
+    // constructor (executed when loading the module)
+    WTModule::WTModule() : Module("RoBMA"){
+      
+      // distributions
+      insert(new DWT1);
+      insert(new DWT2);
+      insert(new DWN1);
+      insert(new DWN2);
+      
+      //effect sizes transformations
+      insert(new d2z);
+      insert(new d2r);
+      insert(new d2logOR);
+      insert(new r2d);
+      insert(new r2z);
+      insert(new r2logOR);
+      insert(new z2r);
+      insert(new z2d);
+      insert(new z2logOR);
+      insert(new logOR2d);
+      insert(new logOR2z);
+      insert(new logOR2r);
+
+      //standard errors transformations
+      insert(new se_d2se_z);
+      insert(new se_d2se_r);
+      insert(new se_d2se_logOR);
+      insert(new se_r2se_d);
+      insert(new se_r2se_z);
+      insert(new se_r2se_logOR);
+      insert(new se_z2se_r);
+      insert(new se_z2se_d);
+      insert(new se_z2se_logOR);
+      insert(new se_logOR2se_d);
+      insert(new se_logOR2se_z);
+      insert(new se_logOR2se_r);
+
+      //prior scaling functions (aproximate linear transformations)
+      insert(new scale_d2z);
+      insert(new scale_d2logOR);
+      insert(new scale_z2d);
+      insert(new scale_z2logOR);
+      insert(new scale_logOR2d);
+      insert(new scale_logOR2z);
     }
-    std::vector<Distribution*> const &dvec = distributions();
-    for (unsigned int i=0;i<dvec.size();++i) {
-      delete dvec[i];
+    
+    // destructor (executed when unloading the module)
+    WTModule::~WTModule() {
+      std::vector<Function*> const &fvec = functions();
+      for (unsigned int i = 0; i < fvec.size(); ++i) {
+        delete fvec[i];
+      }
+      std::vector<Distribution*> const &dvec = distributions();
+      for (unsigned int i=0;i<dvec.size();++i) {
+        delete dvec[i];
+      }
     }
-  }
 
-}
+  }
 }
 
 jags::weightedt::WTModule _weightedt_module;